refactor(node-manager): add task/groups domain dir; rename TaskContextImpl->RunningTaskContext
