+++ packages/node-manager/src/index.ts
@@ -5,8 +5,8 @@
  */
 
 export * from "./ReconcilerBehavior.js";
-export * from "./task/AddNodeToGroup.js";
 export * from "./task/errors.js";
+export * from "./task/groups/AddNodeToGroup.js";
 export * from "./task/Task.js";
 export * from "./task/TaskManagerBehavior.js";
 export * from "./task/TaskRegistry.js";
