+++ packages/node-manager/test/task/TaskContextGateTest.ts
@@ -5,8 +5,8 @@
  */
 
 import { ReconcilerBehavior } from "#ReconcilerBehavior.js";
+import { RunningTaskContext } from "#task/RunningTaskContext.js";
 import { Task } from "#task/Task.js";
-import { TaskContextImpl } from "#task/TaskContextImpl.js";
 import { TaskState } from "#task/types.js";
 import { FakePeer } from "./helpers.js";
 
@@ -24,7 +24,7 @@ function makeContext(peer: FakePeer) {
         task.progress.state = s;
         states.push(s);
     };
-    const ctx = new TaskContextImpl(task, () => peer.asNode(), peer as unknown as ReconcilerBehavior, setState);
+    const ctx = new RunningTaskContext(task, () => peer.asNode(), peer as unknown as ReconcilerBehavior, setState);
     return { task, ctx, states };
 }
